Overview: Redesigned Maps. Details: 1. Added some methods to Map interface. 2. Putted some common methods of maps to base.go. 3. Changed methods of ConcArrMap1 to use base.go. 4. Changed logics of some methods of ConcArrMap1 to make them easier to read and understand sacrificing only constant runtime.
